separate functions and constants
