+++ lib/matrix/format.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getFormatInfo, getVersionInfo } = require('../spec');
+const { getFormatInfo, getVersionInfo } = require('../utils');
 
 const applyFormatInfo = (matrix, eccLevel, maskPattern) => {
   const formatBits = getFormatInfo(eccLevel, maskPattern);
