+++ lib/matrix/index.js
@@ -5,7 +5,7 @@ const { setupPatterns } = require('./patterns');
 const { applyMask, findBestMask } = require('./mask');
 const { applyFormatInfo, applyVersionInfo } = require('./format');
 const { fillData } = require('./data');
-const { getMatrixSize } = require('../spec');
+const { getMatrixSize } = require('../utils');
 
 /**
  * Creates a complete QR matrix from encoded bits
