+++ lib/matrix/patterns.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getAlignmentPatternPositions } = require('../spec');
+const { getAlignmentPatternPositions } = require('../utils');
 
 /**
  * Places a finder pattern at a given position
