+++ lib/spec.js
@@ -160,31 +160,6 @@ const VERSION_INFO = [
   0x27541, 0x28c69,
 ];
 
-const getNsym = (version, eccLevel) => CAPACITIES[version].ecc[eccLevel];
-
-const getCharCountSize = (version, mode) => {
-  if (mode === MODES.BYTE) {
-    return version < 10 ? 8 : 16;
-  }
-  throw new Error(`Mode ${mode} is not supported yet`);
-};
-
-const getDataCapacity = (version, eccLevel) => {
-  const spec = CAPACITIES[version];
-  if (!spec) throw new Error(`Version ${version} is out of range (1-40)`);
-  return spec.total - spec.ecc[eccLevel];
-};
-
-const getMatrixSize = (version) => (version - 1) * 4 + 21;
-
-const getAlignmentPatternPositions = (version) =>
-  ALIGNMENT_POSITIONS[version] || [];
-
-const getFormatInfo = (eccLevel, maskPattern) =>
-  FORMAT_INFO[(eccLevel << 3) | maskPattern];
-
-const getVersionInfo = (version) =>
-  version >= 7 ? VERSION_INFO[version - 7] : null;
 
 // Block structure for each version and ECC level
 // Format: { blocks: [{count, dataWords, eccWords}...] }
@@ -815,12 +790,6 @@ const BLOCK_INFO = {
   },
 };
 
-const getBlockInfo = (version, eccLevel) => {
-  const info = BLOCK_INFO[version];
-  if (!info) throw new Error(`Block info not available for version ${version}`);
-  return info[eccLevel] || [];
-};
-
 module.exports = {
   ECC_LEVELS,
   MODES,
@@ -829,12 +798,4 @@ module.exports = {
   FORMAT_INFO,
   VERSION_INFO,
   BLOCK_INFO,
-  getNsym,
-  getCharCountSize,
-  getDataCapacity,
-  getMatrixSize,
-  getAlignmentPatternPositions,
-  getFormatInfo,
-  getVersionInfo,
-  getBlockInfo,
 };
